+++ src/main/java/org/opensearch/neuralsearch/search/collector/HybridCollectorFactory.java
@@ -18,8 +18,6 @@ import org.opensearch.search.sort.SortAndFormats;
 
 import java.util.Locale;
 
-import static org.opensearch.neuralsearch.settings.NeuralSearchSettings.HYBRID_COLLAPSE_DOCS_PER_GROUP_PER_SUBQUERY;
-
 /**
  * A factory class for creating various types of Hybrid Collectors based on the provided configuration.
  */
@@ -41,10 +39,6 @@ public class HybridCollectorFactory {
         int numHits = hybridCollectorFactoryDTO.getNumHits();
         FieldDoc after = hybridCollectorFactoryDTO.getAfter();
         if (collapseContext != null) {
-            // Collapse is applied
-            int docsPerGroupPerSubquery = HYBRID_COLLAPSE_DOCS_PER_GROUP_PER_SUBQUERY.get(
-                searchContext.indexShard().indexSettings().getSettings()
-            );
             MappedFieldType fieldType = collapseContext.getFieldType();
             if (fieldType instanceof KeywordFieldMapper.KeywordFieldType) {
                 return HybridCollapsingTopDocsCollector.createKeyword(
@@ -52,8 +46,7 @@ public class HybridCollectorFactory {
                     fieldType,
                     sortAndFormats == null ? new Sort(new SortField(null, SortField.Type.SCORE)) : sortAndFormats.sort,
                     numHits,
-                    hitsThresholdChecker,
-                    docsPerGroupPerSubquery
+                    hitsThresholdChecker
                 );
             } else if (fieldType instanceof NumberFieldMapper.NumberFieldType) {
                 return HybridCollapsingTopDocsCollector.createNumeric(
@@ -61,8 +54,7 @@ public class HybridCollectorFactory {
                     fieldType,
                     sortAndFormats == null ? new Sort(new SortField(null, SortField.Type.SCORE)) : sortAndFormats.sort,
                     numHits,
-                    hitsThresholdChecker,
-                    docsPerGroupPerSubquery
+                    hitsThresholdChecker
                 );
             } else {
                 throw new IllegalStateException(